requirements.txt
# Core Framework
fastapi>=0.111.0,<0.112.0
uvicorn[standard]>=0.29.0,<0.30.0
pydantic[email]>=2.7.1,<3.0
pydantic-settings>=2.1.0,<2.3.0
gunicorn>=22.0.0,<23.0.0

# Async & Concurrency
asyncio-contextmanager>=1.0.1
aioredis>=2.0.1,<3.0
asyncpg>=0.29.0,<0.30

# Database & ORM
sqlalchemy>=2.0.23,<3.0
neo4j>=5.20.0,<6.0
psycopg2-binary>=2.9.9,<3.0

# Vector Search
faiss-cpu>=1.8.0,<2.0
numpy>=1.26.0,<2.0

# Caching & Message Queue
redis==5.0.4
python-multipart==0.0.9
celery[redis]==5.3.6
flower==2.0.0

# API & HTTP
httpx>=0.27.0,<0.28
requests>=2.31.0,<3.0

# Core & Security
cryptography==42.0.5
PyJWT>=2.8.0,<3.0
cryptography>=41.0,<43.0
python-jose[cryptography]>=3.3.0,<4.0
passlib[bcrypt]>=1.7.4,<2.0
bcrypt>=4.1.0,<5.0
bcj-python==0.1.1
liboqs-python==0.10.0
py_ecc==6.0.0
authlib>=1.3.0,<2.0

# Observability
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==7.0.0
python-json-logger>=3.0.0
sentry-sdk[fastapi]==2.1.1
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0


# ML & Specialized (Preserved from existing)
transformers==4.40.1
sentence-transformers==3.0.0
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
diffusers==0.27.2
accelerate==0.30.1
openai==1.30.1
groq==0.5.0
together==1.1.0
llama-cpp-python==0.2.76
tiktoken==0.7.0

# LoRA Training (Evolution Engine)
peft>=0.10.0
trl>=0.8.6
bitsandbytes>=0.43.1
datasets>=2.19.0
# Note: unsloth requires specialized installation (see README)

# Cloud & Infrastructure
docker==7.0.0
firebase-admin==6.5.0
google-cloud-firestore==2.15.0
google-cloud-run==0.10.5
google-cloud-tasks==2.16.4
google-cloud-storage==2.16.0
google-cloud-pubsub==2.21.1
gcsfs==2024.3.1
boto3==1.34.106


# Utilities
pyserial==3.5
python-dotenv>=1.0.1,<2.0
tqdm>=4.66.2,<5.0
pandas>=1.5.3,<2.0
slowapi==0.1.9
mtranslate==1.8
moviepy==1.0.3
resend==0.8.0
razorpay==1.4.1
tenacity==8.3.0
PyPDF2==3.0.1
python-docx==1.1.2
tavily-python==0.3.3
pywebpush==2.0.0
itsdangerous==2.1.2
pillow==10.3.0
